Split Data into PID and DataBlock
